Added variable elimination.
Had to change :<: to :>: because, e.g., in the definition of Lex over
two variables [X, Y], X is greater than Y.  This led to the discovery
of a bug: a minimal instance declaration of Ord requires either (<=)
or 'compare', and they are defined by default in terms of each other,
so that if both are omitted (and say only a definition of (>) is
given), then we get an infinite loop trying to compute (<=).  Changed
the definitions of orderings accordingly.